revC2: separate current shunt for the INA233
The original idea in #165 and #135 was to reuse the 0.47 Ohms dropper resistor
we already have for the regulator as current shunt too. As the INA233 has a fixed
measurement range of +- 81.9 mV, this would allow to measure 174 mA. My initial
thought was that this is enough, as the specced load of the TPS73101 is 150 mA.

Review and experimentation has shown that the TPS73101 can be used to 350 mA in
most configurations before limiting set in, for a programmed output of 5 V I
could use it even up to 420 mA. As one of the design goals of using the INA233
is overcurrent protection, the settable range should be higher than what the
regulator can output in best case.

So the design is changed to a 0.15 Ohms current shunt and an additional 0.33 Ohms
dropper resistor in series. This gives a total 0.48 Ohms drop for the regulator
and a measurable current range to 546 mA.

For both resistors a DNPed parallel footprint is provided as before, allowing
the user to reduce the voltage drop if desired.